fix: resolve compilation errors in GeofenceController and PacketDecoder
- GeofenceController: remove .name() call on String (type is already String, not enum)
- PacketDecoder: cast short to int for Integer autoboxing compatibility

+++ backend/src/main/java/com/webinnovation/pingpath/protocol/PacketDecoder.java
@@ -84,8 +84,8 @@ public class PacketDecoder {
                 || protocolNumber == PacketType.LOCATION_4G) {
             if (content.readableBytes() >= 7) {
                 loc.setAccOn(content.readUnsignedByte() == 1);
-                loc.setUploadMode(content.readUnsignedByte());
-                loc.setRealtimeFlag(content.readUnsignedByte());
+                loc.setUploadMode((int) content.readUnsignedByte());
+                loc.setRealtimeFlag((int) content.readUnsignedByte());
                 loc.setMileageMeters(content.readUnsignedInt() * 1000L);  // km → m
             }
         }
